Final commit for HW1, conceding defeat to AWS wonkiness
